Final modelado

+++ UltrAthleticsGen/UltrAthleticsGenNHibernate/CEN/UltrAthletics/ProductoCEN_getTotalValoraciones.cs
@@ -19,7 +19,7 @@ namespace UltrAthleticsGenNHibernate.CEN.UltrAthletics
 {
 public partial class ProductoCEN
 {
-public float GetTotalValoraciones (int p_oid)
+public int GetTotalValoraciones (int p_oid)
 {
         /*PROTECTED REGION ID(UltrAthleticsGenNHibernate.CEN.UltrAthletics_Producto_getTotalValoraciones) ENABLED START*/
